# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api that make authorization",
        "title": "Authorization",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/role": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get all roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Get roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "result page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "quantity roles per page",
                        "name": "qt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all roles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Role"
                            }
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "role params",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RolePartial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "create role successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "400": {
                        "description": "an invalid role param was sent",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "403": {
                        "description": "current user is not admin",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "409": {
                        "description": "role already exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        },
        "/role/{name}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get role by name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Get role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "role name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role return",
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "role does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Delete a role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Delete role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "role name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role deleted",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "403": {
                        "description": "current user is not admin",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "role does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        },
        "/session": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Refresh a user session and set in the response header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Refresh session",
                "responses": {
                    "200": {
                        "description": "user session refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a user session and set in the response header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Create session",
                "parameters": [
                    {
                        "description": "user params",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserSessionPartial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "session created successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "400": {
                        "description": "an invalid user param was sent",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "user does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get all user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "result page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "quantity user per page",
                        "name": "qt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all roles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "user params",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserPartial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "create user successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "400": {
                        "description": "an invalid user param was sent",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "403": {
                        "description": "current user is not admin",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "409": {
                        "description": "username/email already exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get user by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user return",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "user does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Update a user informations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "user params",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserUpdate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update user successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "400": {
                        "description": "an invalid user param was sent",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "403": {
                        "description": "current user is not admin",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "user does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "409": {
                        "description": "username/email already exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Delete a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user deleted",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "403": {
                        "description": "current user is not admin",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "user does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.RolePartial": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserPartial": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "model.UserSessionPartial": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserUpdate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "server.sent": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Session",
            "in": "header"
        }
    }
}